refactor(ux): rename radicleProjectsts -> repo.ts
We don't use the term project any more.
