+++ src/ux/index.ts
@@ -4,5 +4,5 @@ export * from './httpdConnection'
 export * from './patchesView'
 export * from './radCliInstallation'
 export * from './radicleIdentityAuth'
-export * from './radicleProject'
+export * from './repo'
 export * from './settings'
